Disable shared libs when testing with musl-gcc.
After upgrading the testing server to Ubuntu 18.04.5,
I decided to try using /usr/bin/musl-gcc from the distribution
instead of a self-compiled version.

We get a load of errors like this:

    /usr/bin/ld: alloc.o: relocation R_X86_64_32 against `.rodata.str1.8' can not be used when making a shared object; recompile with -fPIC

As far as I can tell, we *are* passing -fPIC when compiling those object
files. Since there were no such errors with the previous setup,
nor have I seeen with any issues with musl / musl-gcc elsewhere,
it may just be a specific issue with the musl-gcc wrapper on
Debian/Ubuntu systems.