Configure with --build --host for i686 builds.
Previously we set CC=i686-linux-gnu-gcc to test i686 builds on a x86-64
host, but configure reports:

    checking build system type... x86_64-pc-linux-gnux32
    checking host system type... x86_64-pc-linux-gnux32

By passing --build=i686-linux-gnu --host=i686-linux-gnu,
configure now detects:

    checking build system type... i686-pc-linux-gnu
    checking host system type... i686-pc-linux-gnu

AFAIK the x32 ABI is experimental and/or not very well supported, so may
be part of the reason for the seemingly random segmentation faults that
we see on the test server. In any case, it's not what we intended to test.